requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.2
frozenlist==1.4.1
future==1.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
python-dateutil==2.9.0.post0
python-whois==0.8.0
pytz==2024.1
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
Werkzeug==3.0.1
yarl==1.9.4
